ecommerce/requirements.txt
asgiref==3.11.1
DateTime==6.0
Django==6.0.3
django-filter==25.2
djangorestframework==3.17.1
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.1
future==1.0.0
PyJWT==1.7.1
pytz==2026.1.post1
serializers==0.2.4
sqlparse==0.5.5
zope.interface==8.2
